MAP-38 - Update Tableau Vogtle Dashboard  (#2)
* Move data generator script to tests directory. Add 'Status of ITAAC Inspections' View to Tableau Dashboard

* Remove old generator directory. Update dashboard progress on MAP-62

* Add calendar data generator. Add visualizations for MAP-38

* Minor visualization updates
