+++ tests/generator/run.sh
